Fixing target selection
All HEADS are loaded, but only requested targets are selected. If an
order is specified by options, targets that appear in the order are
ordered as requested, other targets are then appended as they appeared.
Then this list is used to filter all the available HEADS to create the
display list